requirements.txt
streamlit==1.46.1
flask==2.3.3
groq==0.4.1
requests==2.31.0
flask-cors==4.0.0
python-dateutil==2.8.2
plotly==5.24.1
httpx==0.27.2
python-dotenv==1.1.1
google-generativeai
Pillow
